mcux: sdk-ng: update device/MCX to sdk 26.06.00-pvw1

+++ mcux/mcux-sdk-ng/devices/MCX/MCXW/MCXW236/drivers/power_manager_mcxw23/device/fsl_pm_device.c
@@ -487,7 +487,7 @@ uint64_t PMDEVICE_GetSleepTimer(void)
     }
     return ((uint64_t)cnt * 1000000ULL) / (uint64_t)freq;
 #else
-    uint32_t cnt = OSTIMER_GetCurrentTimerValue(OSTIMER);
+    uint64_t cnt = OSTIMER_GetCurrentTimerValue(OSTIMER);
     uint32_t freq = CLOCK_GetOSTimerClkFreq();
     if (freq == 0U)
     {
